Revert "move packed validation into its own package"
This reverts commit 97f2775db64f2bf7cd1cb2368acf1b74d6777f62.

This complicates the API for not a huge amount of benefit. Instead just
keep the attestation logic in this package (and add apple later).
