+++ example/static/user.html.tmpl
@@ -51,6 +51,7 @@
             <div class="passkey-info-element">
               <div>Attestation object</div>
               <div class="passkey-info-element-data">
+                {{ .AttestationFormat }}<br>
                 {{ .AttestationObject }}
               </div>
             </div>
