+++ webauthn/testdata/blob.jwt
